move switch component to common folder
