+++ bigbluebutton-html5/imports/ui/components/common/switch/styles.js
