+++ bigbluebutton-html5/imports/ui/components/settings/submenus/data-saving/component.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Toggle from '/imports/ui/components/switch/component';
+import Toggle from '/imports/ui/components/common/switch/component';
 import { defineMessages, injectIntl } from 'react-intl';
 import BaseMenu from '../base/component';
 import Styled from './styles';
